New Component::setBounds() method using a RelativeRectangle. Added Justification::appliedToRectangle() and RectanglePlacement::appliedTo(). Removed a behavioural oddity from TopLevelWindow which brought them to the front when made visible. More RelativeRectangle development. AU wrapper now detects app shutdown and closes its UI.

+++ src/audio/audio_sources/juce_ResamplingAudioSource.cpp
@@ -88,7 +88,7 @@ void ResamplingAudioSource::releaseResources()
 
 void ResamplingAudioSource::getNextAudioBlock (const AudioSourceChannelInfo& info)
 {
-    float localRatio;
+    double localRatio;
 
     {
         const ScopedLock sl (ratioLock);
